requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
gunicorn==19.7.1
Markdown==2.6.8
psycopg2==2.7.1
PyJWT==1.5.2
pytz==2017.2
whitenoise==3.3.0
